Crate slosh_lib

Modules

compile

debug

Structs

Reader

Constants

ENV

Env (job control status, etc) for the shell.

SHELL_ENV

Env (job control status, etc) for the shell.

Functions

compile

load_one_expression

new_slosh_vm_with_builtins

new_slosh_vm_with_builtins_and_core

run

run_reader

With the given reader, for each sexp compile then load and execute.

set_builtins

set_builtins_shell

set_initial_load_path

Given a SloshVm and a String, usually the rc file for slosh, set the load-path global variable to facilitate proper loading of scripts.

usage