slosh_lib::debug

Macro file_println

macro_rules! file_println {
    ($filename:expr, $msg:expr) => { ... };
    ($filename:expr, $fmt:expr, $($arg:tt)*) => { ... };
}

A macro similar to println! that writes output to a file in /tmp. Used to experiment but should not end up in committed code, clippy lint expect(unused) helps enforce this.

Examples

file_println!("debug.log", "Hello, world!"); // Writes to a platform specific tmp dir.
file_println!("debug.log", "The value is: {}", 42); // Supports format args.

Note on usage: The expect(unused) fails if macro_export attribute is uncommented because that attribute makes a macro part of the public API of the crate signifying usage, potentially by an external crate.