slosh_lib

Module liner_rules

Structs

LinerWordDividerRule

NewlineForBackSlashAndSlShSyntaxRule

Functions

check_balanced_delimiters 🔒

ensures all provided delimiters are balanced.

check_balanced_delimiters_lisp 🔒

ensures all provided delimiters are balanced for a lisp sexp.

check_balanced_delimiters_shell 🔒

ensures all provided delimiters are balanced for a shell line.

get_liner_words 🔒

is_sexp 🔒

If first non-whitespace char is ‘(’ assume this is an sexp.

make_editor_rules