Crate slosh_lib

Modules

compile

debug

Structs

Reader

Constants

ENV

Env (job control status, etc) for the shell.

SHELL_ENV

Env (job control status, etc) for the shell.

Functions

compile

load_builtins_lisp

load_builtins_lisp_less_sloshrc

load_color

load_core

load_one_expression

load_sloshrc

Usage: (load-rc) | (load-rc “init.slosh)

load_test

new_slosh_vm_with_builtins

run

run_reader

With the given reader, for each sexp compile then load and execute.

run_slosh

set_builtins

All rust based slosh builtins. These functions only prime the vm w/ new functions and do not have any side-effects.

set_builtins_and_shell_builtins

set_environment

Loads the user’s sloshrc file, has side-effects, and sets some important constants in the environment.

set_initial_load_path

Given a SloshVm and a String, usually the rc file for slosh, set the load-path global variable to facilitate proper loading of scripts.

set_shell_builtins

usage